Use explicit int sizes instead of implicit long

+++ lib/Leg/Leg.h
@@ -89,11 +89,11 @@ class Leg {
     float    _tolerance  = 1.0f;
     uint16_t _timeout_ms = 50;
 
-    unsigned long long _last_update  = 0;
-    unsigned long long _last_time    = 0;
-    unsigned long long _now          = 0;
-    float              _delta        = 0.0f;
-    uint16_t           _update_timer = 25;
+    uint64_t _last_update  = 0;
+    uint64_t _last_time    = 0;
+    uint64_t _now          = 0;
+    float    _delta        = 0.0f;
+    uint16_t _update_timer = 25;
 
     void move_joints(vec3_t angles);
     void move_joints(float coxa_angle, float femur_angle, float tibia_angle);
